Move classes to different files
