+++ tests/unit/pipeline/test_routing_batch_function.py
@@ -14,7 +14,7 @@
 
 from typing import List
 
-from distilabel.pipeline.base import _Batch
+from distilabel.pipeline.batch import _Batch
 from distilabel.pipeline.local import Pipeline
 from distilabel.pipeline.routing_batch_function import (
     RoutingBatchFunction,
